Add files via upload
The Sorting Visualizer project is an interactive tool designed to demonstrate the classic Bubble Sort algorithm. This visualizer allows users to observe the sorting process in real-time, providing a clear illustration of how Bubble Sort works. As one of the simplest sorting algorithms, Bubble Sort repetitively steps through a list, compares adjacent elements, and swaps them if they are in the wrong order.Users can customize the array size, value range, and speed of the animation to better understand the algorithm’s efficiency and behavior under various conditions. This project serves as an educational resource for those looking to learn about sorting algorithms and their implementations, making it ideal for students and programming enthusiasts alike. Dive into the world of sorting and enhance your understanding of Bubble Sort with this user-friendly visualizer!